+++ src/main/kotlin/net/pwall/json/schema/subschema/ItemsSchema.kt
@@ -27,7 +27,7 @@ package net.pwall.json.schema.subschema
 
 import java.net.URI
 
-import net.pwall.json.JSONArray
+import net.pwall.json.JSONSequence
 import net.pwall.json.JSONValue
 import net.pwall.json.pointer.JSONPointer
 import net.pwall.json.schema.JSONSchema
@@ -41,7 +41,7 @@ class ItemsSchema(uri: URI?, location: JSONPointer, val itemSchema: JSONSchema) 
 
     override fun validate(json: JSONValue?, instanceLocation: JSONPointer): Boolean {
         val instance = instanceLocation.eval(json)
-        if (instance !is JSONArray)
+        if (instance !is JSONSequence<*>)
             return true
         for (i in instance.indices)
             if (!itemSchema.validate(json, instanceLocation.child(i)))
@@ -52,7 +52,7 @@ class ItemsSchema(uri: URI?, location: JSONPointer, val itemSchema: JSONSchema) 
     override fun validateBasic(relativeLocation: JSONPointer, json: JSONValue?, instanceLocation: JSONPointer):
             BasicOutput {
         val instance = instanceLocation.eval(json)
-        if (instance !is JSONArray)
+        if (instance !is JSONSequence<*>)
             return BasicOutput.trueOutput
         val errors = mutableListOf<BasicErrorEntry>()
         for (i in instance.indices) {
@@ -69,7 +69,7 @@ class ItemsSchema(uri: URI?, location: JSONPointer, val itemSchema: JSONSchema) 
     override fun validateDetailed(relativeLocation: JSONPointer, json: JSONValue?, instanceLocation: JSONPointer):
             DetailedOutput {
         val instance = instanceLocation.eval(json)
-        if (instance !is JSONArray)
+        if (instance !is JSONSequence<*>)
             return createAnnotation(relativeLocation, instanceLocation, "Value is not an array")
         val errors = mutableListOf<DetailedOutput>()
         for (i in instance.indices) {